requirements.txt
streamlit
python-dotenv
swisseph
pytz
fpdf
plotly
google-genai